Fix bug with deleting logos from gallery.
Bug was a result of forgetting that Git does not preserve empty
directories, and hence the `./public/archive` directory into which
deleted logo image files were moved might not exist in a fresh
install.  This fix creates the directory if it does not exist, and
also refactors the code a bit to be more modular and maintainable.